chore(#210): remover campo takeaways (string) — tratar apenas takeawayItems[]
Após Stage 4 (#102, b11e73bf), reviews ficaram com dois campos paralelos
representando takeaways: `takeaways` (string legacy) e `takeawayItems[]`
(array canônico). A migração nunca foi concluída — ReviewToolsPanel
(Extrato) e WeeklyReviewModal (Fila) continuavam escrevendo no string;
StudentReviewsPage renderizava o string como bloco isolado acima do
checklist; WeeklyReviewPage tinha fallback de leitura do string como
sessionNotes; CF createWeeklyReview inicializava `takeaways: null`.

Decisão (Marcio, 30/04/2026): canonizar takeawayItems[]. Conteúdo legado
em prod fica órfão sem migração — se aparecer pro aluno é bug.

Mudanças:

A) Componente compartilhado novo:
   - `src/components/reviews/TakeawaysSection.jsx` (TakeawayItem +
     TakeawaysSection extraídos de WeeklyReviewPage)

B) Hook `useWeeklyReviews`:
   - `appendTakeaway` removido (sem caller após #171)
   - `saveDraftFields` aceita `{sessionNotes,meetingLink,videoLink}` —
     todos undefined-aware (delta vazio = no-op); `takeaways` saiu
   - `closeReview.opts.takeaways` removida

C) `ReviewToolsPanel` (Extrato): Section "Takeaways" textarea →
   `<TakeawaysSection>` consumindo addTakeawayItem/toggle/remove

D) `WeeklyReviewModal` (Fila): tab "Takeaways" textarea →
   `<TakeawaysSection>` (mesma fonte)

E) `StudentReviewsPage`: bloco que renderizava `takeawaysText` removido
   (DOM `data-testid=review-takeaways-text` deixou de existir);
   header do item perdeu marcador "com observações"

F) `WeeklyReviewPage`: fallback `review.takeaways` na inicialização de
   `sessionNotes` removido — agora só sessionNotes vale

G) `reviewUrlValidator`: `validateTakeaways`/`MAX_TAKEAWAYS_LENGTH`
   renomeados para `validateNotesText`/`MAX_NOTES_LENGTH` (validador
   genérico de texto longo, hoje usado só por sessionNotes)

H) CF `createWeeklyReview.js`: removida inicialização `takeaways: null`

Tests:
- `reviewUrlValidator.test.js`: rename refletido
- `useWeeklyReviews.test.js`: closeReview agora valida que takeaways NÃO
  vai no payload + meetingLink chega corretamente
- `StudentReviewsPage.test.jsx`: 4 testes reescritos — agora confirmam
  que campo legado NÃO é renderizado mesmo quando preenchido no doc

Sem migração de dados em prod — campo `takeaways` permanece nos docs
existentes (órfão), simplesmente ignorado pelo código.

Suite: 2743/2743 verde. Lint: baseline preservado.

Closes #210
