+++ docs/dev/issues/issue-210-remove-takeaways-legacy.md
@@ -1,0 +1,69 @@
+# Issue #210 — chore: remover campo takeaways (string), tratar apenas takeawayItems[]
+
+> Branch: `chore/issue-210-remove-takeaways-legacy` · Worktree: `~/projects/issue-210`
+> Versão reservada: v1.49.1 · Lock: CHUNK-08 (escrita)
+
+## Autorização
+
+- [x] Mockup — exceção autorizada (refactor sem UI nova; UX final = TakeawaysSection da WeeklyReviewPage replicada nos outros 2 lugares)
+- [x] Memória de cálculo — exceção autorizada (cleanup arquitetural sem cálculo)
+- [x] Marcio autorizou — 30/04/2026: "ordeno que remova o campo takeaways string! Trate apenas takeawayItems[]. (...) se tiver algo sendo mostrado na revisão que esteja no campo takeaway está errado."
+- [x] Gate Pré-Código liberado
+
+## Context
+
+Após Stage 4 (#102, `b11e73bf`), `students/{uid}/reviews/{rid}` ficou com dois campos paralelos para takeaways: `takeaways` (string legacy) e `takeawayItems[]` (array canônico). A migração nunca foi concluída — `ReviewToolsPanel` (Extrato) e `WeeklyReviewModal` continuam escrevendo no string; `StudentReviewsPage` renderiza o string como bloco isolado acima do checklist; `WeeklyReviewPage` tem fallback de leitura do string como `sessionNotes`; CF `createWeeklyReview` inicializa `takeaways: null`. Sintoma: mentor digita takeaway no Extrato → grava no string → não aparece no checklist do aluno.
+
+Decisão: canonizar `takeawayItems[]`, remover `takeaways` do código. Conteúdo legado em prod fica órfão (sem migração — Marcio: "se aparecer está errado").
+
+## Spec
+
+Ver #210.
+
+## Phases
+
+- A — Componente compartilhado: extrair `TakeawaysSection` + `TakeawayItem` de `WeeklyReviewPage.jsx` para `src/components/reviews/TakeawaysSection.jsx`. Atualizar `WeeklyReviewPage` para importar.
+- B — Hook `useWeeklyReviews`: remover `appendTakeaway`, retirar `takeaways` de `saveDraftFields` e `closeReview.opts`.
+- C — `ReviewToolsPanel`: substituir Section "Takeaways" textarea por `<TakeawaysSection>`.
+- D — `WeeklyReviewModal`: substituir tab "Takeaways" textarea por `<TakeawaysSection>`.
+- E — `StudentReviewsPage`: remover bloco que renderiza `takeawaysText`.
+- F — `WeeklyReviewPage`: remover fallback `review.takeaways` na inicialização de `sessionNotes`.
+- G — Validador: remover `validateTakeaways` + `MAX_TAKEAWAYS_LENGTH` de `reviewUrlValidator.js`.
+- H — CF `createWeeklyReview.js`: remover `takeaways: null` da inicialização.
+- I — Tests adaptados/removidos.
+- J — Smoke browser + PR.
+
+## Sessions
+
+- _(a preencher)_
+
+## Shared Deltas
+
+Aplicados no main na abertura (commit `6ae670de`):
+- `src/version.js` — bump 1.49.0 → 1.49.1 + entrada [RESERVADA]
+- `docs/registry/versions.md` — v1.49.1 reservada
+- `docs/registry/chunks.md` — lock CHUNK-08
+
+Aplicar no encerramento:
+- `docs/registry/chunks.md` — liberar CHUNK-08
+- `docs/registry/versions.md` — marcar 1.49.1 consumida
+- `src/version.js` — finalizar entrada CHANGELOG (sair de [RESERVADA])
+- `docs/PROJECT.md` — bump versão + entrada CHANGELOG
+- `docs/firestore-schema.md` — atualizar reviews para citar `takeawayItems[]` como canônico
+
+## Decisions
+
+- DEC-AUTO-210-01 — `takeawayItems[]` canônico; campo `takeaways` (string) removido sem migração de dados em prod (órfão tolerado).
+
+## Chunks
+
+- CHUNK-08 (Mentor Feedback) — escrita (UI de revisão)
+
+## Aceite
+
+- [ ] `grep -rn "review\.takeaways\b\|review?\.takeaways\b" src/` retorna zero
+- [ ] `grep -rn "appendTakeaway\|validateTakeaways\|MAX_TAKEAWAYS_LENGTH" src/` retorna zero
+- [ ] Suite verde
+- [ ] Mentor adiciona takeaway pelo Extrato → aparece no checklist da WeeklyReviewPage do aluno + PendingTakeaways
+- [ ] Mentor adiciona takeaway pelo Modal → idem
+- [ ] Review legada com `takeaways` string preenchida + `takeawayItems` vazio → bloco NÃO aparece para o aluno
